refactor: Move the useKonnectorsActions hook to Hooks folder
